fix: compute event highlights on client
